secp256k1.cu

+++ secp256k1.cu
@@ -866,14 +866,14 @@ __global__ void test_mod_inverse(uint256_t f, uint256_t g, uint256_t* result) {
 int main() {
     // f = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEFFFFFC2F
     uint256_t f(
-        0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
-        0xFFFFFFFFFFFFFFFFFFFFFFFEFFFFFC2F
+        (__uint128_t)0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
+        (__uint128_t)0xFFFFFFFFFFFFFFFFFFFFFFFEFFFFFC2F
     );
 
     // g = 0x33e7665705359f04f28b88cf897c603c9
     uint256_t g(
-        0x33e7665705359f04ULL,
-        0xf28b88cf897c603c9ULL
+        (__uint128_t)0x33e7665705359f04ULL,
+        (__uint128_t)0xf28b88cf897c603c9ULL
     );
 
     /* g ≡ 1 (mod f): */
@@ -894,3 +894,4 @@ int main() {
     cudaFree(result_device);
     return 0;
 }
+
